Better jQuery encapsulation
Now surrounds the select2.js inclusion with scripts that ensure and
then release (if necessary) window.jQuery, and puts the select2-enabled
version of jQuery into a new namespace to ensure it’s available.

The reason for this is that if this plugin is used along with some
other jQuery-based plugin, we can get multiple versions of jQuery that
overwrite each other.